XIVY-11623 Use new 11.2 API for managed threads and async process start
Instead of starting an own thread with an executor for each event that
is received use the new processStarter API to fire the event
asynchronous. The event bean runtime will then execute the process using
the Ivy internal Job framework.

Instead of starting an own thread with an executor for reading the
events from Kafka use the new API to start a managed event bean thread.

The event bean thread will also be displayed in the Engine Cockpit in
the event bean's detail view.